+++ services/sensor_tracker/src/main.rs
@@ -9,6 +9,7 @@ extern crate serde_derive;
 extern crate uuid;
 
 mod config;
+mod logic;
 mod model;
 mod prawnqtt;
 mod predis;
@@ -21,6 +22,6 @@ fn main() {
 
     let (rx, mqtt_cli) = prawnqtt::start_mqtt(&config);
 
-    predis::receive_updates(rx, &config_clone.to_redis_context(), mqtt_cli);
+    logic::receive_updates(rx, &config_clone.to_redis_context(), mqtt_cli);
     println!("unreachable");
 }
